mkfs.fat: Read geom_start from sysfs
HDIO_GETGEO which provides start member is not supported for loop block
devices. Also start member is only unsigned long so it cannot be used for
bigger disks. So instead read geom_start from sysfs and store it into long
long type to prevent overflow.

+++ src/device_info.h
@@ -34,7 +34,7 @@ struct device_info {
      */
     int geom_heads;
     int geom_sectors;
-    long geom_start;
+    long long geom_start;
     long long geom_size;
 
     /*
